refactor: Move templates to their own crate
This PR logically separates out parts of the code to their own crates. As we've been adding more and more logic into the
template for the Containerfile, compilation times have started to increase. This is mostly due to the fact that askama
is a compilation time templating engine that uses macros to build rust code. I figured the best way to help with that
was to seperate out the templating into its own crate. In order to do that while keeping the CLI as the root package,
I had to also move ops, constants, and the recipe structs into their own sub-packages.
