+++ recipe/Cargo.toml
@@ -1,0 +1,21 @@
+[package]
+name = "blue-build_recipe"
+version.workspace = true
+edition.workspace = true
+description.workspace = true
+repository.workspace = true
+license.workspace = true
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+anyhow.workspace = true
+blue-build_utils = { path = "../utils" }
+chrono.workspace = true
+format_serde_error.workspace = true
+indexmap.workspace = true
+log.workspace = true
+serde.workspace = true
+serde_yaml.workspace = true
+serde_json.workspace = true
+typed-builder.workspace = true
