+++ recipe/src/akmods_info.rs
