+++ recipe/src/lib.rs
@@ -1,0 +1,11 @@
+pub mod akmods_info;
+pub mod image_inspection;
+pub mod module;
+pub mod module_ext;
+pub mod recipe;
+
+pub use akmods_info::*;
+pub use image_inspection::*;
+pub use module::*;
+pub use module_ext::*;
+pub use recipe::*;
