+++ src/lib.rs
@@ -12,8 +12,4 @@
 
 shadow_rs::shadow!(shadow);
 
-pub mod akmods_info;
 pub mod commands;
-pub mod constants;
-pub mod module_recipe;
-mod ops;
