+++ template/Cargo.toml
@@ -1,0 +1,20 @@
+[package]
+name = "blue-build_template"
+version.workspace = true
+edition.workspace = true
+description.workspace = true
+repository.workspace = true
+license.workspace = true
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+askama.workspace = true
+blue-build_utils = { path = "../utils" }
+blue-build_recipe = { path = "../recipe" }
+log.workspace = true
+serde.workspace = true
+serde_yaml.workspace = true
+serde_json.workspace = true
+typed-builder.workspace = true
+uuid.workspace = true
