+++ utils/src/lib.rs
@@ -1,0 +1,209 @@
+pub mod constants;
+
+use std::{
+    ffi::OsStr,
+    fmt::Debug,
+    io::{Error, ErrorKind, Result as IoResult, Write},
+    path::PathBuf,
+    process::{Command, Stdio},
+    time::Instant,
+};
+
+use anyhow::{anyhow, Result};
+use format_serde_error::SerdeError;
+use log::{debug, trace};
+use process_control::{ChildExt, Control};
+use std::{thread, time::Duration};
+
+pub fn check_command_exists(command: &str) -> Result<()> {
+    trace!("check_command_exists({command})");
+    debug!("Checking if {command} exists");
+
+    trace!("which {command}");
+    if Command::new("which")
+        .arg(command)
+        .output()?
+        .status
+        .success()
+    {
+        debug!("Command {command} does exist");
+        Ok(())
+    } else {
+        Err(anyhow!(
+            "Command {command} doesn't exist and is required to build the image"
+        ))
+    }
+}
+
+pub fn append_to_file(file_path: &str, content: &str) -> Result<()> {
+    trace!("append_to_file({file_path}, {content})");
+    debug!("Appending {content} to {file_path}");
+
+    let mut file = std::fs::OpenOptions::new()
+        .append(true)
+        .create(true)
+        .open(file_path)?;
+
+    writeln!(file, "\n{content}")?;
+    Ok(())
+}
+
+pub fn serde_yaml_err(contents: &str) -> impl Fn(serde_yaml::Error) -> SerdeError + '_ {
+    |err: serde_yaml::Error| {
+        let location = err.location();
+        let location = location.as_ref();
+        SerdeError::new(
+            contents.to_string(),
+            (
+                err.into(),
+                location.map_or(0, serde_yaml::Location::line).into(),
+                location.map_or(0, serde_yaml::Location::column).into(),
+            ),
+        )
+    }
+}
+
+pub fn retry<V, F>(mut attempts: u8, delay: u64, f: F) -> anyhow::Result<V>
+where
+    F: Fn() -> anyhow::Result<V>,
+{
+    loop {
+        match f() {
+            Ok(v) => return Ok(v),
+            Err(e) if attempts == 1 => return Err(e),
+            _ => {
+                attempts -= 1;
+                thread::sleep(Duration::from_secs(delay));
+            }
+        };
+    }
+}
+
+#[must_use]
+pub fn home_dir() -> Option<PathBuf> {
+    directories::BaseDirs::new().map(|base_dirs| base_dirs.home_dir().to_path_buf())
+}
+
+// ================================================================================================= //
+// CommandOutput
+// ================================================================================================= //
+
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub struct CommandOutput {
+    pub stdout: String,
+    pub stderr: String,
+}
+
+/// # Attempt to resolve `binary_name` from and creates a new `Command` pointing at it
+/// # This allows executing cmd files on Windows and prevents running executable from cwd on Windows
+/// # This function also initializes std{err,out,in} to protect against processes changing the console mode
+/// #
+/// # Errors
+///
+pub fn create_command<T: AsRef<OsStr>>(binary_name: T) -> IoResult<Command> {
+    let binary_name = binary_name.as_ref();
+    log::trace!("Creating Command for binary {:?}", binary_name);
+
+    let full_path = match which::which(binary_name) {
+        Ok(full_path) => {
+            log::trace!("Using {:?} as {:?}", full_path, binary_name);
+            full_path
+        }
+        Err(error) => {
+            log::trace!("Unable to find {:?} in PATH, {:?}", binary_name, error);
+            return Err(Error::new(ErrorKind::NotFound, error));
+        }
+    };
+
+    #[allow(clippy::disallowed_methods)]
+    let mut cmd = Command::new(full_path);
+    cmd.stderr(Stdio::piped())
+        .stdout(Stdio::piped())
+        .stdin(Stdio::null());
+
+    Ok(cmd)
+}
+
+/// Execute a command and return the output on stdout and stderr if successful
+pub fn exec_cmd<T: AsRef<OsStr> + Debug, U: AsRef<OsStr> + Debug>(
+    cmd: T,
+    args: &[U],
+    time_limit: Duration,
+) -> Option<CommandOutput> {
+    log::trace!("Executing command {:?} with args {:?}", cmd, args);
+    internal_exec_cmd(cmd, args, time_limit)
+}
+
+fn internal_exec_cmd<T: AsRef<OsStr> + Debug, U: AsRef<OsStr> + Debug>(
+    cmd: T,
+    args: &[U],
+    time_limit: Duration,
+) -> Option<CommandOutput> {
+    let mut cmd = create_command(cmd).ok()?;
+    cmd.args(args);
+    exec_timeout(&mut cmd, time_limit)
+}
+
+pub fn exec_timeout(cmd: &mut Command, time_limit: Duration) -> Option<CommandOutput> {
+    let start = Instant::now();
+    let process = match cmd.spawn() {
+        Ok(process) => process,
+        Err(error) => {
+            log::trace!("Unable to run {:?}, {:?}", cmd.get_program(), error);
+            return None;
+        }
+    };
+    match process
+        .controlled_with_output()
+        .time_limit(time_limit)
+        .terminate_for_timeout()
+        .wait()
+    {
+        Ok(Some(output)) => {
+            let stdout_string = match String::from_utf8(output.stdout) {
+                Ok(stdout) => stdout,
+                Err(error) => {
+                    log::warn!("Unable to decode stdout: {:?}", error);
+                    return None;
+                }
+            };
+            let stderr_string = match String::from_utf8(output.stderr) {
+                Ok(stderr) => stderr,
+                Err(error) => {
+                    log::warn!("Unable to decode stderr: {:?}", error);
+                    return None;
+                }
+            };
+
+            log::trace!(
+                "stdout: {:?}, stderr: {:?}, exit code: \"{:?}\", took {:?}",
+                stdout_string,
+                stderr_string,
+                output.status.code(),
+                start.elapsed()
+            );
+
+            if !output.status.success() {
+                return None;
+            }
+
+            Some(CommandOutput {
+                stdout: stdout_string,
+                stderr: stderr_string,
+            })
+        }
+        Ok(None) => {
+            log::warn!("Executing command {:?} timed out.", cmd.get_program());
+            log::warn!("You can set command_timeout in your config to a higher value to allow longer-running commands to keep executing.");
+            None
+        }
+        Err(error) => {
+            log::trace!(
+                "Executing command {:?} failed by: {:?}",
+                cmd.get_program(),
+                error
+            );
+            None
+        }
+    }
+}
